Staff can make unavailable dates available

+++ src/frontend/src/components/BookingOverview/BookingListComponent.jsx
@@ -18,6 +18,7 @@ function BookingListComponent() {
     const [timeSlots, setTimeSlots] = useState([]);
     const [date, setDate] = useState(moment(new Date()).format('YYYY-MM-DD'));
     const [closeDayConfirmation, setCloseDayConfirmation] = useState(true);
+    const [openDayConfirmation, setOpenDayConfirmation] = useState(true);
     const [showCreateModal, setShowCreateModal] = useState(false);
     let modalBooking = {
         bookingID: 0,
@@ -62,23 +63,47 @@ function BookingListComponent() {
     }
     
     const startCloseDayConfirmation = () => setCloseDayConfirmation(false);
+    const startOpenDayConfirmation = () => setOpenDayConfirmation(false);
 
-    //Boka upp alla timeslots för denna dag
     function confirmCloseDay(bookingDate) {
         bookingDate = moment(bookingDate).format('YYYY-MM-DD')
 
-        BookingDataService.getBookingsByDate(bookingDate).then(response => {
-            if (response.data.length === 0){
-                BookingDataService.deleteBookingTimes(bookingDate).then(
-                    () => {
+        BookingDataService.getTimeSlotsByDate(bookingDate).then(response => {
+            if(response.data.length !== 0) {
+                BookingDataService.getBookingsByDate(bookingDate).then(response => {
+                    if (response.data.length === 0){
+                        BookingDataService.deleteBookingTimes(bookingDate).then(
+                            () => {
+                                window.location.reload();
+                            });
+                    } else {
+                        alert("Det finns bokningar inlagda för denna dag. Ta bort bokningarna innan du stänger av dagen.")
                         window.location.reload();
-                    });
+                    }
+                })
             } else {
-                alert("Det finns bokningar inlagda för denna dag. Ta bort bokningarna innan du stänger av dagen.")
+                alert("Denna dag är redan stängd.")
                 window.location.reload();
             }
         })
 
+
+    }
+
+    function confirmOpenDay(bookingDate) {
+        bookingDate = moment(bookingDate).format('YYYY-MM-DD')
+
+        BookingDataService.getTimeSlotsByDate(bookingDate).then(response => {
+            if (response.data.length === 0){
+                BookingDataService.addBookingTimes(bookingDate).then(
+                    () => {
+                        window.location.reload();
+                    });
+                } else {
+                    alert("Denna dag är redan öppen.")
+                    window.location.reload();
+                }
+        })
     }
 
 
@@ -86,12 +111,16 @@ function BookingListComponent() {
         let prevDate = new Date(date)
         prevDate.setDate(prevDate.getDate() - 1)
         changeDate(moment(prevDate).format('YYYY-MM-DD'))
+        setCloseDayConfirmation(true)
+        setOpenDayConfirmation(true)
     }
 
     function onNextDate() {
         let nextDate = new Date(date)
         nextDate.setDate(nextDate.getDate() + 1)
         changeDate(moment(nextDate).format('YYYY-MM-DD'))
+        setCloseDayConfirmation(true)
+        setOpenDayConfirmation(true)
     }
 
     const changeDate = (newDate) => {
@@ -170,6 +199,12 @@ function BookingListComponent() {
                         Stäng dag
                     </Button> : <Button variant="danger" onClick={() => confirmCloseDay(date)}>
                         Är du säker på att du vill stänga av bokningar för denna dag?
+                    </Button>}
+                    
+                    {openDayConfirmation ? <Button variant="primary" onClick={startOpenDayConfirmation}>
+                        Öppna dag
+                    </Button> : <Button variant="primary" onClick={() => confirmOpenDay(date)}>
+                        Är du säker på att du vill öppna bokningar för denna dag?
                     </Button>}
                 </div>
